+++ src/client.ts
@@ -1,4 +1,3 @@
-import { AxiosInstance } from 'axios';
 import { AuthParams, ClientParams, ParamsOauth1, ParamsOauth2 } from './types';
 import { version as packageVersion } from './version';
 
@@ -14,6 +13,7 @@ import {
   dispatchersService,
   paymentsService,
   profilesService,
+  notificationsService,
 } from './services';
 
 import {
@@ -23,6 +23,7 @@ import {
   createOAuth2HttpClient,
 } from './http';
 import { validateConfig } from './utils';
+import { HttpInstance } from './http/types';
 
 export interface OAuth1Authenticate {
   /**
@@ -106,7 +107,7 @@ type Authenticate<
 > = T extends ParamsOauth1 ? OAuth1Authenticate : OAuth2Authenticate;
 
 export interface Client<T extends ClientParams> {
-  raw: AxiosInstance;
+  raw: HttpInstance;
   /**
    * The template service manages templates used to build emails. It can be used to retrieve, create, update or delete templates as well as resolving them.
    * @see https://developers.extrahorizon.io/services/templates-service/1.0.13/
@@ -152,6 +153,11 @@ export interface Client<T extends ClientParams> {
    * @see https://developers.extrahorizon.io/services/profiles-service/1.1.3/
    */
   profiles: ReturnType<typeof profilesService>;
+  /**
+   * A service that handles push notifications.
+   * @see https://developers.extrahorizon.io/services/notifications-service/1.0.8/
+   */
+  notifications: ReturnType<typeof notificationsService>;
   /**
    * The user service stands in for managing users themselves, as well as roles related to users and groups of users.
    * @see https://developers.extrahorizon.io/services/users-service/1.1.7/
@@ -226,6 +232,7 @@ export function createClient<T extends ClientParams>(rawConfig: T): Client<T> {
     dispatchers: dispatchersService(httpWithAuth),
     payments: paymentsService(httpWithAuth),
     profiles: profilesService(httpWithAuth),
+    notifications: notificationsService(httpWithAuth),
     auth: {
       ...authService(httpWithAuth),
       authenticate: (oauth: AuthParams): Promise<void> =>