+++ src/http/oauth2.ts
@@ -133,5 +133,12 @@ export function createOAuth2HttpClient(
     setTokenData(tokenResult.data);
   }
 
-  return Object.assign(httpWithAuth, { authenticate, confirmMfa });
+  return {
+    ...httpWithAuth,
+    authenticate,
+    confirmMfa,
+    get userId() {
+      return tokenData?.userId;
+    },
+  };
 }