+++ src/services/index.ts
@@ -9,3 +9,4 @@ export * from './mails';
 export * from './dispatchers';
 export * from './payments';
 export * from './profiles';
+export * from './notifications';